session4/requirements.txt
streamlit
streamlit_lottie
streamlit_option_menu
streamlit_mermaid
pandas
matplotlib
numpy==1.26.1
scikit-learn
seaborn
tensorflow
requests
tabulate
networkx
altair 
plotly 
pillow
graphviz
networkx
boto3